Implemented validators, comfirmation questions for deleting.

+++ AlexGuitarsShop/ViewModels/GuitarViewModel.cs
@@ -4,7 +4,7 @@ public class GuitarViewModel
 {
     public int Id { get; init; }
     public string Name { get; init; }
-    public ushort Price { get; init; }
+    public int Price { get; init; }
     public string Description { get; init; }
     public IFormFile Avatar { get; init; }
     public byte[] Image { get; set; }
